# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Digital Life Admin
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTaskBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  priority?: number;
  status?: string;
  /** @nullable */
  deadline?: Date | null;
  tags?: string[];
  /** @nullable */
  estimatedMinutes?: number | null;
}